Use a baseof template with a 'main' block
Splitting the opening and closing html tags across partial templates
(in this case header and footer) is a bad idea.
Instead, use a base template with a main block for content.

We reuse the header.html and footer.html partial templates for the
body content header and footer.

We also introduce header-css.html for the included css files (to make
the base template a little bit smaller/readable).