Drop theme in favor of site local files
We do not plan on publishing the theme, or using multiple/different
themes. So we can drop the theme configuration and folder overhead in
favor of site-local files.
(This also prevents possible ambiguity between site-local files and theme
files where the theme would win, but separate places to edit would be
unnecessarily spread out.)