Layout: Drop footer copyright notice
A copyright notice is not necessary to receive and retain copyright.

A copyright year is typically/originally added so users may see when the
copyright expires.
However, nobody will wait for copyright on our website to expire to make
use of its content.
And if necessary, the source repository retains a history.

I also do not see how the notice could be useful when saving or printing
the website. I doubt that’s a typical use case, and the above
argumentation applies.

I am also wondering if always printing the current year is correct.
Does the copyright start when we distribute/publish, even when the
content does not change?