Security: Fix branch bundle using list template
The list template was introduced to fix a warning.
But that changed the behavior of the conflicting leaf bundle with the same
name as the security branch bundle; and changed the effectively visible
page from the leaf to the branch bundle using the list template.

This change makes it an explicit leaf bundle with the content intended for
it.