Add release redirects based on the info in version.json
Currently, the stable release links are hardcoded and the snapshot release ones are retrieved every hour (since the start of the webserver) with a regex formula by reading Apache's directory index at https://dl.mumble.info.

This commit introduces a new system, which adds the proper redirects based on the info in version.json.

version.json is our new file containing info about the current stable and snapshot release for all platforms; it's a replacement for snapshot.xml.

Changes to the file are detected in real time thanks to fsnotify. The watcher runs in a dedicated thread.

Redirects are stored in a map and handled by a dedicated function. This is necessary because http.ServeMux doesn't have a function to remove an handler and adding a new one for the same pattern results in a panic.

Concurrent access (e.g. the handler reading the map and the watcher writing it) is protected with a sync.Mutex.