Slice: Add generated slice HTML documentation 1.2.0, 1.2.7, 1.3.0
Up to now we host our slice documentation on http://mumble.sourceforge.net/slice/.

However, having this additional host and service separately makes no sense
other than for historic reasons; keeping the old URLs in tact.

With more documentation moving to the website, it makes sense to integrate
the slice documentation into the website as well.

1.2.19 still has the same slice definition as 1.2.7, so no generating
necessary for that.
The old site on mumble.sf.net also hosts the generated HTML for 1.2.0,
1.2.7 and (snapshot) 1.3.0.
The 1.3.0 content added here is definitely of stable version 1.3.0.

Fixes #60

This documentation is generated with slice2html from Ice 3.6. Ice 3.7 no
longer provides this binary, and instead refers to doxygen which
generates complexer HTML. This means we will have to evaluate and
decide on how we want to handle this in the future. See #75.