CI: Use Ubuntu-latest instead of specific version
Microsoft currently provides two hosted Ubuntu agents; version 16.04 and 18.04.
This change effectively changes 16.04 to 18.04 being used. But in the future we will automatically build on a newer agent and Ubuntu version once it becomes available.

Using a specific version makes sense for constant, reproducible builds.
However, we build products that do not depend on libraries.
The Hugo website produces static files, and the webserver produces a binary without dependencies.
We can use latest to stay on edge and not have to worry about upgrading and the build breaking when an old image and agent is dropped.

We use the default Go installation now and expect it to remain backwards compatible.
We use a specific Hugo version we install which is independent of this process. But we would expect it to remain backwards compatible as well.
We do not expect problematic hidden differences from the image being updated; from Ubuntu or Go being updated.
This will drop our need to think about and manage this infrastructure, build environment work - at least for now and for as long as those expectations are not broken.
We trade the need of conscious analytical upgrading to the danger of subtle and (initially) unnoticed issues.