FIX: Publish dates of blog entries rendered incorrectly
The date that was listed on the website did not match the date that was
specified in the actual blog post markdown files. The issue was an
invalid dateformat string used in our template.

Apparently Go (hugo?) has super weird requirements for that format
string. The month has to be January, the day must be 02 (was 01 in our
case) and the year must be 2006.
Ref: https://github.com/htr3n/hyde-hyde/issues/2#issuecomment-377405853