Use generated Table of Contents on Contribute page
The manually added ToC targets anchors. As a result when updating headlines the ToC must also be updated.
This is error prone.

The resulting ToC list changes slightly in which headlines it displays.
Instead of a selection over three levels deep, the first two headline levels are listed.
This works fine for the content at hand.