Home: Show latest news and release post
Both have validity and are important, a post may be published in both categories in close time.
We want to promote both kinds.

Latest release remains valid and interesting to users landing on home.
Latest news is also a good entry and nudge that we have them even if older.