Sponsors: Adjust text on Cyberia
I think the previous text was not very neutral or professional.
I found reading the text quite irritating and somewhat confusing.

I suggest we change the previous hype-tone, praise- and promise-tone to a more professional tone.

It still describes the collective, the issue solved, and the sponsored item, as well as our appreciation.
It also IMO reduces preknowledge and unrealistic or inappropriate expectation towards the reader (which was my main issue with the tone).