FIX: REF_NOT_FOUND error
For some reason when locally hosting the website using hugo server it
would result in a REF_NOT_FOUND error for that relref. However it seems
like this is running fine in the deployed version of the site.

This might be an issue of which hugo version is being used but in any
case removing the trailing slash seems to resolve the error and yield
the desired result.