Create 2025-12-27-finding-foss-bugs.md
This is a request to add a guest blog post to the mumble blog, as invited by the developers.  

I worked with @krzmbl after identifying the issue in #6875 on the main mumble repo, and this is the story of how I found the bug.

The goal of this post is to encourage people to find ways to contribute to open source even if they're not specifically software developers.